ready to run

+++ src/main.rs
@@ -1,9 +1,8 @@
-use discord::model::{Event, ReactionEmoji, EmojiId};
+use discord::model::{Event, ReactionEmoji};
 use discord::Discord;
 use std::env;
 use urlencoding::encode;
 use std::fs::File;
-use std::io::Write;
 use std::path::Path;
 use downloader::{Downloader, Download};
 
@@ -19,10 +18,17 @@ static REACTIONS: [&'static str; 9] = [
     "\u{1f1f8}"
 ];
 
+static mut EQUATION_INDEX: usize = 0;
+
 fn main() {
+
     let file_contents= std::fs::read_to_string("equations.txt")
         .expect("Something went wrong reading the file");
     let equations: Vec<_> = file_contents.lines().collect();
+
+    unsafe {
+        EQUATION_INDEX = rand::random::<usize>() % equations.len();
+    }
 
     let mut name = load_equation(&equations);
 
@@ -43,7 +49,7 @@ fn main() {
                     let response = discord.send_file(
                         message.channel_id,
                         &format!("two plus two is four, minus one that's {}", name),
-                        std::fs::File::open("equation.png").unwrap(),
+                        File::open("equation.png").unwrap(),
                         "equation.png"
                     ).unwrap();
                     for reaction in REACTIONS {
@@ -67,16 +73,16 @@ fn main() {
 }
 
 fn load_equation<'a>(equations: &Vec<&'a str>) -> &'a str {
-    static mut EQUATION_INDEX: usize = 0;
 
-    std::fs::remove_file("equation.png").unwrap();
+    if Path::new("equation.png").exists() {
+        std::fs::remove_file("equation.png").unwrap();
+    }
 
     let (name, equation) = unsafe { equations[EQUATION_INDEX].split_once(" | ").unwrap() };
-    dbg!(equation);
     let url = format!("https://latex.codecogs.com/png.latex?\\dpi{{150}}&space;\\fn_cm&space;\\LARGE&space;{}", encode(equation));
 
     let download = Download::new(&*url).file_name("equation.png".as_ref());
-    downloader::Downloader::builder().build().unwrap().download(&[download]).expect("download failed");
+    Downloader::builder().build().unwrap().download(&[download]).expect("download failed");
 
     unsafe {
         EQUATION_INDEX += 1;
